requirements.txt
discord.py[voice]
yt_dlp
PyNaCl
ffmpeg-python
spotipy
aiohttp
python-dotenv
